+++ principal/estruturas.cpp
@@ -2,15 +2,15 @@ typedef struct bomb {
 	//clock de início de cada bomba
 	clock_t start[9];
 	//total de bombas disponíveis
-	int total;
+	short int total;
 	//bombas no tabuleiro
-	int inboard;
+	short int inboard;
 	//coordenadas de cada bomba
 	coord co[9];
 	//frame que será executado para cada bomba
-	int framenumber[9];
+	short int framenumber[9];
 	//potência do fogo
-	int fire;
+	short int fire;
 	//slot usado;
 	bool used[9];
 };
@@ -19,9 +19,9 @@ typedef struct bomb {
 
 typedef struct monster {
 	//número de vidas de cada monstro
-	int life[10];
-	int total;
-	int inboard;
+	short int life[10];
+	short int total;
+	short int inboard;
 	//tipo de monstro (1-7)
 	char type[10];
 	coord co[10];
@@ -30,7 +30,7 @@ typedef struct monster {
 //===============================================
 
 typedef struct hero {
-	int life;
+	short int life;
 	coord co;
 	//cor do bomberball
 	short int color;
